refactor: move code from files to a new model for db management
