+++ includes/model/Openedx_Woocommerce_Plugin_Enrollment.php
@@ -20,6 +20,16 @@ class Openedx_Woocommerce_Plugin_Enrollment {
     public $post_type = 'openedx_enrollment';
 
     /**
+     * The variable to store the log manager object.
+     * 
+     * @var     object
+     * @access  private
+     * @since   1.1.1
+     */
+
+    private $log_manager;
+
+    /**
      * Constructor function.
      *
      * @access  public
@@ -29,8 +39,18 @@ class Openedx_Woocommerce_Plugin_Enrollment {
     public function __construct( $parent ) {
         $this->parent = $parent;
 
-        // Register all the model related hooks
+        // Register all the model related hooks and log manager.
         $this->register_hook_callbacks();
+        $this->register_log_manager();
+    }
+
+    /**
+     * Register the log manager object to the variable.
+     * 
+     * @return void
+     */
+    public function register_log_manager() {
+        $this->log_manager = new Openedx_Woocommerce_Plugin_Log();
     }
 
     protected function register_hook_callbacks() {
@@ -275,53 +295,7 @@ class Openedx_Woocommerce_Plugin_Enrollment {
             $this->updatePost($post_id, 'enrollment-pending');
         }
 
-
-        $this->createChangeLog($post_id, $old_data_array, $enrollment_arr, $enrollment_action);
-    }
-
-
-    /**
-     * Creates a new log entry in the database.
-     * 
-     * @param int $post_id The post ID.
-     * @param array $old_data_array The old data array.
-     * @param array $enrollment_arr The new data array.
-     * @param string $enrollment_action The API action to perform.
-     * 
-     * @return void
-     */
-    public function createChangeLog( $post_id, $old_data_array, $enrollment_arr, $enrollment_action ) {
-
-        global $wpdb;
-        $tabla_logs = $wpdb->prefix . 'enrollment_logs_req_table';
-        $new_post = get_post($post_id);
-    
-        $date = current_time('mysql', true);
-        $user_id = get_current_user_id(); 
-        $username = get_user_by('id', $user_id)->user_login; 
-    
-        if (empty($old_data_array['enrollment_course_id'])) {
-            $enrollment_action = 'Object Created';
-        } 
-    
-        if (empty($old_data_array['enrollment_course_id'])) {
-            $old_data_array = '--';
-        }
-    
-        $log_data = array(
-            'post_id' => $post_id,
-            'mod_date' => $date,
-            'user' => $username,
-            'action_name' => $enrollment_action,
-            'object_before' => json_encode($old_data_array),
-            'object_after' => json_encode($enrollment_arr)
-        );
-    
-        $wpdb->insert($tabla_logs, $log_data);
-    
-        if ($wpdb->last_error) {
-            error_log('There was an error generating a register: ' . $wpdb->last_error);
-        }
+        $this->log_manager->createChangeLog($post_id, $old_data_array, $enrollment_arr, $enrollment_action);
     }
 
     /**
